add name transformation to general utils script
