+++ app/utils/general.py
@@ -126,5 +126,13 @@ def to_csv(df) -> bytes:
     output.seek(0)
     return output.getvalue()
 
+###############################################################################
 
-
+def transform_name(name):
+    # Re-formats legislator names
+    parts = re.split(r"\s+", name.replace(",", ""))
+    no_title = len(name.split(",")) == 1
+    if no_title:
+        return f"{parts[-1]}, {' '.join(parts[0:-1])}"
+    else:
+        return f"{parts[-2]}, {' '.join(parts[0:-2])}, {parts[-1].replace(',', '')}"
